refactor: 🔄 create deep-required cases in separate file
